Allow to import Gtk from gaphor.ui
So that versions are properly defined.
